+++ activerecord/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Support `:if` and `:unless` options in `has_secure_token`
+
+    *Griffin Smith*
+
 *   Use `version` column as primary key for schema_migrations table because
     schema_migrations versions are guaranteed to be unique.